Move as classes relacionadas com AssetsInReport de workarea.py para reports_maker.py
