+++ src/scielo/bin/xml/prodtools/data/workarea.py
@@ -451,129 +451,3 @@ class DocumentOutputFiles(object):
                   self.style_report_filename, self.pmc_dtd_report_filename,
                   self.pmc_style_report_filename, self.ctrl_filename]:
             fs_utils.delete_file_or_folder(f)
-
-
-def AssetsInReport(pkg_path,
-                       acron=None, issue_label=None,
-                       serial_path=None, web_app_path=None,
-                       web_url=None):
-    if serial_path:
-        return CollectionAssetsInReport(
-                    pkg_path, acron, issue_label,
-                    serial_path, web_app_path, web_url)
-    return BasicAssetsInReport(pkg_path)
-
-
-class BasicAssetsInReport(object):
-
-    def __init__(self, pkg_path):
-        self.pkg_path = pkg_path
-
-    @property
-    def result_path(self):
-        return os.path.dirname(self.pkg_path)
-
-    @property
-    def img_path(self):
-        return self.pkg_path
-
-    @property
-    def pdf_path(self):
-        return self.pkg_path
-
-    @property
-    def xml_path(self):
-        return self.pkg_path
-
-    @property
-    def report_path(self):
-        return os.path.join(self.result_path, 'errors')
-
-    @property
-    def report_link(self):
-        return self.report_path
-
-    @property
-    def img_link(self):
-        return self.img_path
-
-    @property
-    def pdf_link(self):
-        return self.pdf_path
-
-    @property
-    def xml_link(self):
-        return self.xml_path
-
-
-class CollectionAssetsInReport(BasicAssetsInReport):
-
-    def __init__(self, pkg_path, acron, issue_label,
-                 serial_path, web_app_path, web_url):
-        super().__init__(pkg_path)
-        self.acron = acron
-        self.issue_path = os.path.join(acron, issue_label)
-        self._img_revistas_subdir = os.path.join(
-            'img', 'revistas', self.issue_path)
-        self._pdf_subdir = os.path.join('pdf', self.issue_path)
-        self._xml_subdir = os.path.join('xml', self.issue_path)
-        self._reports_subdir = os.path.join('reports', self.issue_path)
-        self.serial_path = serial_path
-        self.web_app_path = web_app_path
-        self.web_url = web_url
-
-    @property
-    def result_path(self):
-        return os.path.join(self.serial_path, self.issue_path)
-
-    @property
-    def img_path(self):
-        return os.path.join(
-            self.web_app_path, "htdocs", self._img_revistas_subdir)
-
-    @property
-    def pdf_path(self):
-        return os.path.join(self.web_app_path, "bases", self._pdf_subdir)
-
-    @property
-    def xml_path(self):
-        return os.path.join(self.web_app_path, "bases", self._xml_subdir)
-
-    @property
-    def report_path(self):
-        if self.web_url:
-            return os.path.join(
-                self.web_app_path, 'htdocs', self._reports_subdir)
-        return self.serial_report_path
-
-    @property
-    def report_link(self):
-        if self.web_url:
-            return self.web_url + "/" + self._reports_subdir
-        return self.serial_report_path
-
-    @property
-    def serial_report_path(self):
-        return os.path.join(self.result_path, 'base_xml', 'base_reports')
-
-    @property
-    def serial_base_xml_path(self):
-        return os.path.join(self.result_path, 'base_xml', 'base_source')
-
-    @property
-    def img_link(self):
-        if self.web_url:
-            return self.web_url + "/" + self._img_revistas_subdir
-        return self.img_path
-
-    @property
-    def pdf_link(self):
-        if self.web_url:
-            return self.web_url + "/" + self._pdf_subdir
-        return self.pdf_path
-
-    @property
-    def xml_link(self):
-        if self.web_url:
-            return self.web_url + "/" + self._reports_subdir
-        return self.xml_path
